requirements.txt
torch==1.12.1+cu116 --extra-index-url https://download.pytorch.org/whl/cu116
torchvision==0.13.1+cu116 --extra-index-url https://download.pytorch.org/whl/cu116
torchaudio==0.12.1 --extra-index-url https://download.pytorch.org/whl/cu116

numpy
pandas
scikit-learn
scikit-image
ipython
Pillow
matplotlib
seaborn
gdown
tqdm
SimpleITK

git+https://github.com/qubvel/segmentation_models.pytorch
segmentation-models-pytorch
